Cannon Locomotion Done
Chris and Andrew: Cannon is complete and all we need is to do the main shooting part... and then insert a target to destroy. Game is horrible looking (aesthetically) but is working relatively smooth as of now.

To do next time:
Work on cannon shooting, make the wall fall ;), update the main menu with textures, and add UI to the cannon ammunition.

Co-Authored-By: Andrew Munoz <amunoz247@users.noreply.github.com>